chore: continue nav bar button refactor
